+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Prepare campusonline-api cache
+
 ## v0.1.8
 
 - implement request caching of users representing the current persons of the current result